TSTFIX: Fix ordereddict tests
