+++ .travis.yml
@@ -14,9 +14,9 @@ python:
 env:
     - ENV="python=2.6 numpy=1.6"
     - ENV="python=2.7 numpy"
-    - ENV="python=3.2"
     - ENV="python=3.3 numpy"
     - ENV="python=3.4 numpy"
+    - ENV="python=3.5 numpy"
 
 virtualenv:
     system_site_packages: true
@@ -29,11 +29,11 @@ before_install:
     # Python 3.2 is not supported by Miniconda, so we use the package manager for that run.
     # NumPy has a bug in python 3 that is only fixed in the latest version,
     #  hence the below wget of numpy/_import_tools.py from github.
-    - if [[ $ENV == python=3.2 ]]; then
+    - if [[ $ENV == python=3.3* ]]; then
         sudo apt-get install python3-numpy;
         wget https://raw.githubusercontent.com/numpy/numpy/master/numpy/_import_tools.py -O /home/travis/virtualenv/python3.2_with_system_site_packages/lib/python3.2/site-packages/numpy/_import_tools.py;
         sudo apt-get install python3-scipy;
-        travis_retry pip install cython flake8 six;
+        travis_retry pip install cython flake8 six nose networkx matplotlib;
       else
         wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh;
         bash miniconda.sh -b -p $HOME/miniconda;
@@ -50,6 +50,9 @@ before_install:
 install:
     - tools/header.py "Dependency versions"
     - tools/build_versions.py
+    - if [[ $ENV == python=2.6* ]]; then
+        travis_retry conda install unittest2;
+      fi
     - export PYTHONWARNINGS=all
     - python setup.py build_ext --inplace
