+++ skfuzzy/control/tests/test_ordereddict.py
@@ -9,20 +9,20 @@ import copy
 from random import shuffle
 from collections import MutableMapping
 
-from skfuzzy.control.ordereddict import OrderedDict
 try:
-    import _mapping_tests as mapping_tests
+    from collections import OrderedDict
 except ImportError:
-    # python 2.6
-    pass
+    from skfuzzy.control.ordereddict import OrderedDict
+
+import _mapping_tests as mapping_tests
 import sys
 version = sys.version_info[:2]
 
 from numpy.testing.decorators import skipif
+from _skipclass import skipclassif
 
 
 class TestOrderedDict(unittest.TestCase):
-    @skipif(version != (2, 7))
     def test_init(self):
         with self.assertRaises(TypeError):
             OrderedDict([('a', 1), ('b', 2)], None)                                 # too many args
@@ -45,7 +45,6 @@ class TestOrderedDict(unittest.TestCase):
         self.assertEqual(list(d.items()),
             [('a', 1), ('b', 2), ('c', 3), ('d', 4), ('e', 5), ('f', 6), ('g', 7)])
 
-    @skipif(version != (2, 7))
     def test_update(self):
         with self.assertRaises(TypeError):
             OrderedDict().update([('a', 1), ('b', 2)], None)                        # too many args
@@ -86,12 +85,10 @@ class TestOrderedDict(unittest.TestCase):
         self.assertEqual(list(d.items()),
             [('a', 1), ('b', 2), ('c', 3), ('d', 4), ('e', 5), ('f', 6), ('g', 7)])
 
-    @skipif(version != (2, 7))
     def test_abc(self):
         self.assertIsInstance(OrderedDict(), MutableMapping)
         self.assertTrue(issubclass(OrderedDict, MutableMapping))
 
-    @skipif(version != (2, 7))
     def test_clear(self):
         pairs = [('c', 1), ('b', 2), ('a', 3), ('d', 4), ('e', 5), ('f', 6)]
         shuffle(pairs)
@@ -100,7 +97,6 @@ class TestOrderedDict(unittest.TestCase):
         od.clear()
         self.assertEqual(len(od), 0)
 
-    @skipif(version != (2, 7))
     def test_delitem(self):
         pairs = [('c', 1), ('b', 2), ('a', 3), ('d', 4), ('e', 5), ('f', 6)]
         od = OrderedDict(pairs)
@@ -110,7 +106,6 @@ class TestOrderedDict(unittest.TestCase):
             del od['a']
         self.assertEqual(list(od.items()), pairs[:2] + pairs[3:])
 
-    @skipif(version != (2, 7))
     def test_setitem(self):
         od = OrderedDict([('d', 1), ('b', 2), ('c', 3), ('a', 4), ('e', 5)])
         od['c'] = 10           # existing element
@@ -118,22 +113,20 @@ class TestOrderedDict(unittest.TestCase):
         self.assertEqual(list(od.items()),
                          [('d', 1), ('b', 2), ('c', 10), ('a', 4), ('e', 5), ('f', 20)])
 
-    @skipif(version != (2, 7))
     def test_iterators(self):
         pairs = [('c', 1), ('b', 2), ('a', 3), ('d', 4), ('e', 5), ('f', 6)]
         shuffle(pairs)
         od = OrderedDict(pairs)
         self.assertEqual(list(od), [t[0] for t in pairs])
-        self.assertEqual(od.keys()[:], [t[0] for t in pairs])
-        self.assertEqual(od.values()[:], [t[1] for t in pairs])
-        self.assertEqual(od.items()[:], pairs)
-        self.assertEqual(list(od.iterkeys()), [t[0] for t in pairs])
-        self.assertEqual(list(od.itervalues()), [t[1] for t in pairs])
-        self.assertEqual(list(od.iteritems()), pairs)
+        # self.assertEqual(od.keys()[:], [t[0] for t in pairs])
+        # self.assertEqual(od.values()[:], [t[1] for t in pairs])
+        # self.assertEqual(od.items()[:], pairs)
+        self.assertEqual(list(od.keys()), [t[0] for t in pairs])
+        self.assertEqual(list(od.values()), [t[1] for t in pairs])
+        self.assertEqual(list(od.items()), pairs)
         self.assertEqual(list(reversed(od)),
                          [t[0] for t in reversed(pairs)])
 
-    @skipif(version != (2, 7))
     def test_popitem(self):
         pairs = [('c', 1), ('b', 2), ('a', 3), ('d', 4), ('e', 5), ('f', 6)]
         shuffle(pairs)
@@ -144,7 +137,6 @@ class TestOrderedDict(unittest.TestCase):
             od.popitem()
         self.assertEqual(len(od), 0)
 
-    @skipif(version != (2, 7))
     def test_pop(self):
         pairs = [('c', 1), ('b', 2), ('a', 3), ('d', 4), ('e', 5), ('f', 6)]
         shuffle(pairs)
@@ -169,7 +161,6 @@ class TestOrderedDict(unittest.TestCase):
         with self.assertRaises(KeyError):
             m.pop('a')
 
-    @skipif(version != (2, 7))
     def test_equality(self):
         pairs = [('c', 1), ('b', 2), ('a', 3), ('d', 4), ('e', 5), ('f', 6)]
         shuffle(pairs)
@@ -185,7 +176,6 @@ class TestOrderedDict(unittest.TestCase):
         # different length implied inequality
         self.assertNotEqual(od1, OrderedDict(pairs[:-1]))
 
-    @skipif(version != (2, 7))
     def test_copying(self):
         # Check that ordered dicts are copyable, deepcopyable, picklable,
         # and have a repr/eval round-trip
@@ -210,7 +200,6 @@ class TestOrderedDict(unittest.TestCase):
             self.assertEqual(len(dup), len(od))
             self.assertEqual(type(dup), type(od))
 
-    @skipif(version != (2, 7))
     def test_yaml_linkage(self):
         # Verify that __reduce__ is setup in a way that supports PyYAML's dump() feature.
         # In yaml, lists are native but tuples are not.
@@ -220,18 +209,6 @@ class TestOrderedDict(unittest.TestCase):
         # '!!python/object/apply:__main__.OrderedDict\n- - [a, 1]\n  - [b, 2]\n'
         self.assertTrue(all(type(pair)==list for pair in od.__reduce__()[1]))
 
-    @skipif(version != (2, 7))
-    def test_reduce_not_too_fat(self):
-        # do not save instance dictionary if not needed
-        pairs = [('c', 1), ('b', 2), ('a', 3), ('d', 4), ('e', 5), ('f', 6)]
-        od = OrderedDict(pairs)
-        self.assertEqual(len(od.__reduce__()), 2)
-        # cyordereddict: remove this test because extension types don't allow
-        # assigning arbitrary attributes
-        # od.x = 10
-        # self.assertEqual(len(od.__reduce__()), 3)
-
-    @skipif(version != (2, 7))
     def test_repr(self):
         od = OrderedDict([('c', 1), ('b', 2), ('a', 3), ('d', 4), ('e', 5), ('f', 6)])
         self.assertEqual(repr(od),
@@ -239,7 +216,6 @@ class TestOrderedDict(unittest.TestCase):
         self.assertEqual(eval(repr(od)), od)
         self.assertEqual(repr(OrderedDict()), "OrderedDict()")
 
-    @skipif(version != (2, 7))
     def test_repr_recursive(self):
         # See issue #9826
         od = OrderedDict.fromkeys('abc')
@@ -247,7 +223,6 @@ class TestOrderedDict(unittest.TestCase):
         self.assertEqual(repr(od),
             "OrderedDict([('a', None), ('b', None), ('c', None), ('x', ...)])")
 
-    @skipif(version != (2, 7))
     def test_setdefault(self):
         pairs = [('c', 1), ('b', 2), ('a', 3), ('d', 4), ('e', 5), ('f', 6)]
         shuffle(pairs)
@@ -266,7 +241,6 @@ class TestOrderedDict(unittest.TestCase):
                 return 0
         self.assertEqual(Missing().setdefault(5, 9), 9)
 
-    @skipif(version != (2, 7))
     def test_reinsert(self):
         # Given insert a, insert b, delete a, re-insert a,
         # verify that a is now later than b.
@@ -277,15 +251,13 @@ class TestOrderedDict(unittest.TestCase):
         od['a'] = 1
         self.assertEqual(list(od.items()), [('b', 2), ('a', 1)])
 
-    @skipif(version != (2, 7))
     def test_views(self):
         s = 'the quick brown fox jumped over a lazy dog yesterday before dawn'.split()
         od = OrderedDict.fromkeys(s)
-        self.assertEqual(list(od.viewkeys()), s)
-        self.assertEqual(list(od.viewvalues()), [None for k in s])
-        self.assertEqual(list(od.viewitems()), [(k, None) for k in s])
+        self.assertEqual(list(od.keys()), s)
+        self.assertEqual(list(od.values()), [None for k in s])
+        self.assertEqual(list(od.items()), [(k, None) for k in s])
 
-    @skipif(version != (2, 7))
     def test_override_update(self):
         # Verify that subclasses can override update() without breaking __init__()
         class MyOD(OrderedDict):
@@ -294,23 +266,27 @@ class TestOrderedDict(unittest.TestCase):
         items = [('a', 1), ('c', 3), ('b', 2)]
         self.assertEqual(list(MyOD(items).items()), items)
 
-if sys.version_info[:2] > (2, 6):
 
-    class GeneralMappingTests(mapping_tests.BasicTestMappingProtocol):
-        type2test = OrderedDict
+@skipclassif(sys.version_info[:2] > (2, 6), "Python version above 2.6")
+class GeneralMappingTests(mapping_tests.BasicTestMappingProtocol):
+    type2test = OrderedDict
 
-        @skipif(version != (2, 7))
-        def test_popitem(self):
-            d = self._empty_mapping()
-            self.assertRaises(KeyError, d.popitem)
+    @skipif(sys.version_info[:2] > (2, 6))
+    def test_popitem(self):
+        d = self._empty_mapping()
+        self.assertRaises(KeyError, d.popitem)
 
-    class MyOrderedDict(OrderedDict):
-        pass
 
-    class SubclassMappingTests(mapping_tests.BasicTestMappingProtocol):
-        type2test = MyOrderedDict
+@skipclassif(sys.version_info[:2] > (2, 6), "Python version above 2.6")
+class MyOrderedDict(OrderedDict):
+    pass
 
-        @skipif(version != (2, 7))
-        def test_popitem(self):
-            d = self._empty_mapping()
-            self.assertRaises(KeyError, d.popitem)
+
+@skipclassif(sys.version_info[:2] > (2, 6), "Python version above 2.6")
+class SubclassMappingTests(mapping_tests.BasicTestMappingProtocol):
+    type2test = MyOrderedDict
+
+    @skipif(sys.version_info[:2] > (2, 6))
+    def test_popitem(self):
+        d = self._empty_mapping()
+        self.assertRaises(KeyError, d.popitem)
